refactor: move instrumentation into internal
